refactor: tighten seed lookback typing and drop unused cycle settings param
SEED_HISTORY_LOOKBACK_DAYS is now a Long so both sync services pass it
directly to Instant.minus; API_VERSION stays a string in the Kraken catalog;
and performCycleWithStableSession no longer takes an unused settings
parameter.

+++ src/main/kotlin/com/gemini/krakenbot/service/impl/history/LedgersSyncService.kt
@@ -76,7 +76,7 @@ class LedgersSyncService(
         // re-fetched and deduplicated rather than missed. Unseeded initial sync and recovery both
         // bound to SEED_HISTORY_LOOKBACK_DAYS like TradeHistorySyncService to avoid fetching
         // historical entries that would be immediately pruned.
-        val seedBound = nowProvider().minus(PrecisionConstants.SEED_HISTORY_LOOKBACK_DAYS.toLong(), ChronoUnit.DAYS)
+        val seedBound = nowProvider().minus(PrecisionConstants.SEED_HISTORY_LOOKBACK_DAYS, ChronoUnit.DAYS)
         val startSec = effectiveLatest?.minusSeconds(300)?.epochSecond
         val isRecoveringInitialSync = !isSeeded && readInitialPaginationOffset() != null
         val paginationStartSec = if (isRecoveringInitialSync) {
